footer fix & bootstrap footer
footer was fixed to page before, but this would get in the way of  the other text. Changed it so that it's more responsive by either being fixed to bottom of the page or trailing a little after the main content.

pretty much copy and pasted the tutorial footer from react bootstrap. Got the icons to work. Will have to scrap a lot of the tutorial bits and change the design. Mostly used for something to look at for now